+++ src/App.tsx
@@ -1,58 +1,16 @@
 import "./App.css";
-import {
-  BrowserRouter,
-  Routes,
-  Route,
-  Navigate,
-  useLocation,
-} from "react-router-dom";
-import { Header } from "./components/Header/header";
-import AllPokemonsPage from "./pages/all-pokemons";
-import MyPokemonsPage from "./pages/my-pokemons";
-import ArenaPage from "./pages/arena-page";
-import { Pokador } from "./assets/catch-button";
+import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import LoginPage from "./pages/login";
-
-const AppContent = () => {
-  const location = useLocation();
-  const items = [
-    {
-      name: "All Pokémons",
-      href: "/all-pokemons",
-      isActive: location.pathname === "/all-pokemons",
-    },
-    {
-      name: (
-        <span className="flex items-center gap-2">
-          My Pokémons
-          <span className="inline-flex items-center justify-center w-5 h-5">
-            <Pokador size={18} />
-          </span>
-        </span>
-      ),
-      href: "/my-pokemons",
-      isActive: location.pathname === "/my-pokemons",
-    },
-  ];
-
-  return (
-    <>
-      <Header items={items} />
-      <Routes>
-        <Route path="/login" element={<LoginPage />} />
-        <Route path="/all-pokemons" element={<AllPokemonsPage />} />
-        <Route path="/my-pokemons" element={<MyPokemonsPage />} />
-        <Route path="/arena" element={<ArenaPage />} />
-        <Route path="*" element={<Navigate to="/all-pokemons" replace />} />
-      </Routes>
-    </>
-  );
-};
+import HomePage from "./pages/home";
 
 function App() {
   return (
     <BrowserRouter>
-      <AppContent />
+      <Routes>
+        <Route path="/login" element={<LoginPage />} />
+        <Route path="/*" element={<HomePage />} />
+        <Route path="*" element={<Navigate to="/" replace />} />
+      </Routes>
     </BrowserRouter>
   );
 }
